update: add endpoint update product

+++ internal/dto/product-create-update-request.go
@@ -9,7 +9,7 @@ type ProductCreateUpdateRequest struct {
 	Price             int64  `validate:"required" json:"price"`
 	Stock             int    `validate:"required" json:"stock"`
 	Wholesaler        string `validate:"required" json:"wholesaler"`
-	ShippingCost      int    `validate:"required" json:"shipping_cost"`
+	ShippingCost      int64  `validate:"required" json:"shipping_cost"`
 	ShippingInsurance int    `validate:"required" json:"shipping_insurance"`
 	Conditions        string `validate:"required" json:"condition"`
 	PreOrder          string `validate:"required" json:"pre_order"`
